refactor: Correcting file names and moving them to ui
